Framework improvisation
