--- automated_demo_project/src/pages/DavidJeremiah.ts
@@ -1,365 +1,0 @@
-import { Page, Locator, expect } from '@playwright/test';
-import LoginLocators from "../Locators/LoginPageLocators.json"
-
-export class CombinedPages {
-  
-  async waitForTime(arg0: number) {
-      await this.page.waitForTimeout(arg0);
-  }
-  async verifySuccessfulLogin() {
-      const myaccountTab = this.page.locator('#tab_my-account');
-      await expect(myaccountTab).toBeVisible();
-  }
-  protected page: Page;
-  private readonly baseUrl: string = 'https://www.davidjeremiah.org/';
-
-  // Element locators - organized by page type
-  private readonly splashElements = {
-    skipButton: () => this.page.locator('[data-testid="skip-button"]')
-      .or(this.page.locator('text="Skip"'))
-      .or(this.page.locator('button:has-text("Skip")'))
-      .or(this.page.locator('text="SKIP AND GO TO DAVIDJEREMIAH.ORG"'))
-  };
-
-  private readonly homePageElements = {
-    promiseOfHeavenImage: () => this.page.locator('img[alt*="Promise of Heaven"], img[title*="Promise of Heaven"], a[title*="Promise of Heaven"]').first(),
-    loginLink: () => this.page.getByText('Login', { exact: true }),
-    preorderTodayButton: () => this.page.locator('button:has-text("Preorder Today!"), a:has-text("Preorder Today!")')
-  };
-
-  private readonly bookDetailsElements = {
-    preorderButton: () => this.page.locator('button:has-text("Preorder Now"), a:has-text("Preorder Now")').first()
-  };
-
-  private readonly loginElements = {
-    emailInput: () => this.page.locator('input[type="email"], input[name="email"], input[id="email"]').first(),
-    passwordInput: () => this.page.locator('input[type="password"], input[name="password"], input[id="password"]').first(),
-    loginButton: () => this.page.locator('button[type="submit"], button:has-text("Login"), button:has-text("Log In"), input[type="submit"]').first()
-  };
-
-  constructor(page: Page) {
-    this.page = page;
-  }
-
-  // ================================
-  // BASE PAGE FUNCTIONALITY
-  // ================================
-
-  async navigateToUrl(url?: string): Promise<void> {
-    const targetUrl = url || this.baseUrl;
-    await this.page.goto(targetUrl);
-  }
-
-  async waitForPageLoad(timeout: number = 30000): Promise<void> {
-    await this.page.waitForLoadState('domcontentloaded', { timeout });
-  }
-
-  async waitForElement(locator: Locator, timeout: number = 30000): Promise<void> {
-    await locator.waitFor({ state: 'visible', timeout });
-  }
-
-  async clickElement(locator: Locator): Promise<void> {
-    await this.waitForElement(locator);
-    await locator.click();
-  }
-
-  async fillInput(locator: Locator, text: string): Promise<void> {
-    await this.waitForElement(locator);
-    await locator.fill(text);
-  }
-
-  async getText(locator: Locator): Promise<string> {
-    await this.waitForElement(locator);
-    return await locator.textContent() || '';
-  }
-
-  async getTitle(): Promise<string> {
-    return await this.page.title();
-  }
-
-  async getCurrentUrl(): Promise<string> {
-    return this.page.url();
-  }
-
-  async takeScreenshot(name: string): Promise<Buffer | null> {
-    try {
-      if (this.page.isClosed()) {
-        console.log(`Cannot take screenshot '${name}' - page is closed`);
-        return null;
-      }
-      return await this.page.screenshot({ path: `screenshots/${name}.png` });
-    } catch (error: any) {
-      console.log(`Failed to take screenshot '${name}':`, error.message);
-      return null;
-    }
-  }
-
-  async scrollToElement(locator: Locator): Promise<void> {
-    await locator.scrollIntoViewIfNeeded();
-  }
-
-  async isElementVisible(locator: Locator): Promise<boolean> {
-    try {
-      await locator.waitFor({ state: 'visible', timeout: 5000 });
-      return true;
-    } catch {
-      return false;
-    }
-  }
-
-  async waitForUrl(urlPattern: string | RegExp, timeout: number = 30000): Promise<void> {
-    await this.page.waitForURL(urlPattern, { timeout });
-  }
-
-  // ================================
-  // SPLASH PAGE FUNCTIONALITY
-  // ================================
-
-  async performSplashPageActions(): Promise<void> {
-    await this.page.goto(`${this.baseUrl}`);
-    // await this.page.waitForTimeout(15000);
-    await this.page.waitForLoadState('domcontentloaded');
-    // Wait for potential animations
-    await this.waitForPageLoad();
-    
-    // Try to click skip button if present
-    try {
-      await this.clickElement(this.splashElements.skipButton());
-      await this.waitForPageLoad();
-    } catch (error) {
-      console.log('Skip button not found, may already be on main site');
-    }
-  }
-
-  async clickSkipButton(): Promise<void> {
-    await this.clickElement(this.splashElements.skipButton());
-  }
-
-  // ================================
-  // HOMEPAGE FUNCTIONALITY
-  // ================================
-
-  async performHomePageNavigation(): Promise<void> {
-    await this.navigateToUrl();
-    await this.waitForPageLoad();
-  }
-
-  async clickPromiseOfHeavenImage(): Promise<void> {
-    await this.clickElement(this.homePageElements.promiseOfHeavenImage());
-  }
-
-  async clickLoginLink(): Promise<void> {
-    await this.clickElement(this.homePageElements.loginLink());
-  }
-
-  async clickPreorderTodayButton(): Promise<void> {
-    await this.clickElement(this.homePageElements.preorderTodayButton());
-  }
-
-  async verifyHomePage(): Promise<void> {
-    await expect(this.page).toHaveURL(/davidjeremiah\.org/);
-    await expect(this.page).toHaveTitle(/DavidJeremiah/);
-  }
-
-  async performHomePageVerification(): Promise<void> {
-    // Verify we're on the correct URL
-    const currentUrl = await this.getCurrentUrl();
-    expect(currentUrl).toContain('davidjeremiah.org');
-    
-    // Verify basic page elements are visible
-    const bodyElement = this.page.locator('body');
-    await expect(bodyElement).toBeVisible();
-    
-    // Check for header (try multiple selectors)
-    const headerSelectors = ['.site-header', 'header', '#header', '.header'];
-    let headerFound = false;
-    for (const selector of headerSelectors) {
-      try {
-        const headerElement = this.page.locator(selector).first();
-        await expect(headerElement).toBeVisible({ timeout: 5000 });
-        headerFound = true;
-        break;
-      } catch (error) {
-        continue;
-      }
-    }
-    if (!headerFound) {
-      console.log('No standard header found, but continuing with test...');
-    }
-    
-    await this.waitForPageLoad();
-  }
-
-  // ================================
-  // BOOK DETAILS FUNCTIONALITY
-  // ================================
-
-  async clickPreorderButton(): Promise<void> {
-    await this.clickElement(this.bookDetailsElements.preorderButton());
-  }
-
-  async clickPreorderNowButton(): Promise<void> {
-    await this.clickElement(this.bookDetailsElements.preorderButton());
-  }
-
-  async performBookInteraction(): Promise<void> {
-    try {
-      // Try to click "Preorder Today!" button first
-      await this.clickPreorderTodayButton();
-      await this.waitForPageLoad();
-    } catch (error) {
-      console.log('Preorder Today button not found, continuing...');
-    }
-    
-    try {
-      // Click the Promise of Heaven image
-      await this.clickPromiseOfHeavenImage();
-      await this.waitForPageLoad();
-      
-      // Click the "Preorder Now" button
-      await this.clickPreorderNowButton();
-      await this.waitForPageLoad();
-    } catch (error) {
-      console.log('Promise of Heaven interaction failed, continuing with workflow...');
-    }
-  }
-
-  // ================================
-  // LOGIN FUNCTIONALITY
-  // ================================
-
-  async login(email: string, password: string): Promise<void> {
-    await this.fillInput(this.loginElements.emailInput(), email);
-    await this.fillInput(this.loginElements.passwordInput(), password);
-    await this.clickElement(this.loginElements.loginButton());
-  }
-
-  async verifyLoginPage(): Promise<void> {
-    await expect(this.loginElements.emailInput()).toBeVisible();
-    await expect(this.loginElements.passwordInput()).toBeVisible();
-    await expect(this.loginElements.loginButton()).toBeVisible();
-  }
-  async verifyInvalidLoginError(): Promise<void> {
-    await expect(this.page.locator(LoginLocators.InvalidLoginErrorMsg.locator)).toBeVisible();
-  }
-
-  async performLoginProcess(email: string, password: string): Promise<void> {
-    try {
-      // Click the Login link
-      await this.clickLoginLink();
-      await this.waitForPageLoad();
-      
-      // Take screenshot before login attempt
-      await this.takeScreenshot('before-login');
-      
-      // Attempt login with provided credentials - with shorter timeout
-      try {
-        await this.fillInput(this.loginElements.emailInput(), email);
-        await this.fillInput(this.loginElements.passwordInput(), password);
-        
-        // Try to click login button with shorter timeout
-        const loginButton = this.page.locator("//span[text()='Login']").click()
-        
-        
-      } catch (loginError: any) {
-        console.log('Login button not found or login failed:', loginError.message);
-        // Continue anyway since login might not be working due to invalid credentials
-      }
-      
-      await this.waitForPageLoad();
-      await this.page.waitForTimeout(2000);
-      
-      // Verify we're still on the David Jeremiah site (if page still exists)
-      if (!this.page.isClosed()) {
-        const currentUrl = await this.getCurrentUrl();
-        expect(currentUrl).toContain('davidjeremiah.org');
-        
-        // Take a screenshot for verification
-        await this.takeScreenshot('automation-complete');
-      }
-      
-    } catch (error: any) {
-      console.log('Login step encountered an issue:', error.message);
-      
-      // Try to take a screenshot if page is still available
-      await this.takeScreenshot('login-error');
-    }
-    
-    console.log('Login step completed (verification skipped due to invalid test credentials)');
-  }
-
-  // ================================
-  // COMBINED WORKFLOW METHODS
-  // ================================
-
-  async performCompleteWorkflow(email: string = 'dkassenova@davidjeremiah.org', password: string = 'LaMesa123!'): Promise<void> {
-    try {
-      // Step 1: Handle splash page
-      await this.performSplashPageActions();
-      
-      // Step 2: Verify homepage
-      await this.performHomePageVerification();
-      
-      // Step 3: Interact with book content
-      await this.performBookInteraction();
-      
-      // Step 4: Return to main page
-      await this.navigateToUrl();
-      await this.waitForPageLoad();
-      
-      // Try to verify homepage but don't fail if it doesn't work
-      try {
-        await this.verifyHomePage();
-      } catch (error) {
-        console.log('Homepage verification failed, but continuing...');
-      }
-      
-      // Step 5: Perform login process
-      await this.performLoginProcess(email, password);
-      
-      // Step 6: Final verification (only if page still exists)
-      if (!this.page.isClosed()) {
-        const currentUrl = await this.getCurrentUrl();
-        expect(currentUrl).toContain('davidjeremiah.org');
-        await this.takeScreenshot('workflow-completed');
-      }
-      
-      console.log('✅ Complete automation workflow finished successfully!');
-      
-    } catch (error: any) {
-      console.log('Workflow encountered an error:', error.message);
-      
-      // Try to take error screenshot if page still exists
-      if (!this.page.isClosed()) {
-        await this.takeScreenshot('workflow-error');
-      }
-      
-      // Re-throw the error so test fails appropriately
-      throw error;
-    }
-  }
-
-  async performSimpleNavigation(): Promise<void> {
-    // Navigate directly to homepage
-    await this.navigateToUrl();
-    await this.waitForPageLoad();
-    
-    // Verify homepage
-    await this.performHomePageVerification();
-    
-    // Verify page title (flexible matching)
-    try {
-      const title = await this.getTitle();
-      expect(title).toContain('David Jeremiah');
-    } catch (error) {
-      console.log('Title verification failed, but continuing...');
-    }
-    
-    // Take a screenshot for documentation
-    await this.takeScreenshot('homepage-navigation-test');
-    
-    // Verify current URL is correct
-    const currentUrl = await this.getCurrentUrl();
-    expect(currentUrl).toContain('davidjeremiah.org');
-  }
-}
