+++ automated_demo_project/src/pages/HomePage.ts
@@ -1,10 +1,10 @@
-import { HomeSteps } from '../steps/HomeSteps';
 import HomepageLocators from '../Locators/HomePageLocators.json';
-import { expect, Page, Locator } from '@playwright/test';
+import { expect, Page } from '@playwright/test';
+import { BasePage } from './BasePage';
 
-export class HomePage {
+export class HomePage extends BasePage {
  async assertSubscriptionErrorMsg() {
-    // await this.page.locator(HomepageLocators.SubscriptionErrorMsg.locator).scrollIntoViewIfNeeded();
+   
    await expect(this.page.locator(HomepageLocators.SubscriptionErrorMsg.locator)).toContainText("Please enter a valid email address");
    console.log(await this.page.locator(HomepageLocators.SubscriptionErrorMsg.locator).textContent());
 
@@ -13,7 +13,7 @@ export class HomePage {
    await this.page.fill(HomepageLocators.SubscriptionInput.locator, arg0);
  }
  async assertSubscriptionSuccessMsg() {
-    // await this.page.locator(HomepageLocators.SubscriptionSuccessMsg.locator).scrollIntoViewIfNeeded();
+    
    await expect(this.page.locator(HomepageLocators.SubscriptionSuccessMsg.locator)).toContainText("You've been successfully signed up!");
  }
  async SelectedCurrency(arg0: string) {
@@ -38,8 +38,8 @@ export class HomePage {
       await this.page.waitForTimeout(4000);
     }
   }
-  constructor(private page: Page) {
-    this.page = page;
+  constructor( page: Page) {
+  super(page);
   }
   async addToCart() {
     await this.page.click(HomepageLocators.addToCart.locator);
@@ -88,5 +88,42 @@ export class HomePage {
   async assertSelectedCurrency(text:string)
   {
     await expect(this.page.locator(HomepageLocators.SelectedCurrency.locator)).toHaveAttribute('title', text);
+  }
+    async performHomePageVerification(): Promise<void> {
+    // Verify we're on the correct URL
+    const currentUrl = await this.getCurrentUrl();
+    expect(currentUrl).toContain('davidjeremiah.org');
+    
+    // Verify basic page elements are visible
+    const bodyElement = this.page.locator('body');
+    await expect(bodyElement).toBeVisible();
+    
+    // Check for header (try multiple selectors)
+    const headerSelectors = ['.site-header', 'header', '#header', '.header'];
+    let headerFound = false;
+    for (const selector of headerSelectors) {
+      try {
+        const headerElement = this.page.locator(selector).first();
+        await expect(headerElement).toBeVisible({ timeout: 5000 });
+        headerFound = true;
+        break;
+      } catch (error) {
+        continue;
+      }
+    }
+    if (!headerFound) {
+      console.log('No standard header found, but continuing with test...');
+    }
+    
+    await this.waitForPageLoad();
+  }
+
+   async performHomePageNavigation(): Promise<void> {
+    await this.navigateToUrl();
+    await this.waitForPageLoad();
+  }
+  async verifyHomePage(): Promise<void> {
+    await expect(this.page).toHaveURL(/davidjeremiah\.org/);
+    await expect(this.page).toHaveTitle(/DavidJeremiah/);
   }
 }
